Pulling auth test code so it can be reused.
